+++ source/entities/pawns/Bullet.cpp
@@ -154,7 +154,7 @@ void Bullet::TickUpdate(const double deltaTime)
 
 unsigned int Bullet::GetDamage() const { return _calibre.damage; }
 
-double Bullet::GetDamageRadius() const { return _calibre.damageRadius; }
+float Bullet::GetDamageRadius() const { return _calibre.damageRadius; }
 
 std::string Bullet::GetAuthor() const { return _author; }
